pyproject.toml
[project]
name = "django-shop-final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-debug-toolbar>=6.0.0",
    "django-extensions>=4.1",
    "factory-boy>=3.3.3",
    "faker>=37.6.0",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pandas-stubs>=2.3.2.250827",
    "psycopg[binary]>=3.2.9",
    "tqdm>=4.67.1",
]
